Assignments/04-A04/candyAPI/requirements.txt
pymongo
rich
fastapi
uvicorn
gunicorn
pillow
io